backend/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.4
python-socketio==5.11.3
python-multipart==0.0.12
pydantic==2.9.1
motor==3.6.0
PyJWT==2.9.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
